requirements.txt
gunicorn==20.0.4
orjson==3.2.2
Flask==1.1.2
PyYAML==5.3.1
pyjwt
Flask-SQLAlchemy
pymysql
PyJWT